lib-requirements.txt
Adafruit-Blinka==6.13.0
Adafruit_DHT==1.4.0
RPi.GPIO==0.7.0
Flask==2.0.1
configparser==5.0.2